+++ tests/test_callbacks.py
@@ -151,10 +151,10 @@ def test_keras_convert_sequential():
     print(wandb_model_out)
     assert wandb_model_out == {'_type': 'graph', 'format': 'keras',
                                'nodes': [
-                                   {'name': 'dense_1', 'id': 'dense_1', 'class_name': 'Dense',
-                                    'output_shape': (None, 4), 'num_parameters': 16},
                                    {'name': 'dense_1_input', 'id': 'dense_1_input', 'class_name': 'InputLayer',
                                     'output_shape': (None, 3), 'num_parameters': 0},
+                                   {'name': 'dense_1', 'id': 'dense_1', 'class_name': 'Dense',
+                                    'output_shape': (None, 4), 'num_parameters': 16},
                                    {'name': 'dense_2', 'id': 'dense_2', 'class_name': 'Dense',
                                     'output_shape': (None, 5), 'num_parameters': 25},
                                    {'name': 'dense_3', 'id': 'dense_3', 'class_name': 'Dense', 'output_shape': (None, 6), 'num_parameters': 36}],